Limit doubling time to smoothed line
